Refactor: Move Minesweeper to its own page and update tsconfig
