--- next.config.ts
@@ -1,13 +1,0 @@
-const nextConfig = {
-  async redirects() {
-    return [
-      {
-        source: "/",
-        destination: "/minesweeper",
-        permanent: false,
-      },
-    ];
-  },
-};
-
-export default nextConfig;
